prevent race conditions in learner assignation, refactor for readability and documentation
